requirements.txt
python-telegram-bot[job-queue]==21.0
psutil